### 2018-12-01 #1114 @whedon
[PRE REVIEW]: rdbsp: An Open Source Rock Dynamic Behavior Simulation Program
**Submitting author:** @zhangning737 (<a href="http://orcid.org/0000-0003-3597-0691">Zhang Ning</a>)
**Repository:** <a href="https://github.com/zhangning737/rdbsp.git" target ="_blank">https://github.com/zhangning737/rdbsp.git</a>
**Version:** v1.0.0
