### 2018-12-03 #1114 @danielskatz
👋  @zhangning737 - there is going to need to be quite a bit more information in the README.  For example, I was going to use that to assign an editor, which I can't do without having some idea what this software does.  Some of the content from the paper probably needs to be copied into the README.  You also might want to look at the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) (see documentation) to see what else you should be adding.
