### 2018-12-04 #1114 @lheagy
Thanks @danielskatz: I don't have much experience with C / C++ , but I would be happy to be the editor on this submission
