### 2018-12-04 #1114 @lheagy
👋 Hi @zhangning737, thanks for your submission to JOSS. Before we send this out for review, please have a look at the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). Specifically,
- [ ] an [OSI approved license](https://opensource.org/licenses/alphabetical) needs to be included in the repository
- [ ] documentation for the software should be provided. This should include
    - [ ] a statement of need
    - [ ] example usage
    - [ ] API documentation
    - [ ] community guidelines
    - [ ] functionality
    - [ ] tests
- [ ] the README should include a high level overview of the documentation that includes

Please have a look as some of the other accepted JOSS submissions, e.g.:
- [Ising_OPV v4.0: Experimental Tomography Data Import, Interpretation, and Analysis](http://joss.theoj.org/papers/0b6b9d572386a1520d3c6eff4563d22c)
- [Lithology: A Landlab submodule for spatially variable rock properties](http://joss.theoj.org/papers/74487c5a6820fb2fe2898960ad6d2ea0)

For now, I will tag this submission as pending enhancements. Please keep me posted on your progress and once these items are addressed, we can work on finding reviewers.
