### 2019-02-06 a3c303d
Add documentation and README.md to repo
