### 2019-02-06 #1114 @zhangning737
@lheagy update the repo with documentation and READ.md. pls check it. I will those qualified. Any else needed?
