### 2019-02-10 #1114 @lheagy
Thanks for your updates @zhangning737, there are still a few things that I did not see:
- [ ]  an explanation of how to run an example
- [ ] an explanation of how to run the tests. Note that the tests need to be distinct from the example usage. Tests should provide some objective way to make sure that the code is functioning as expected.
- [ ] a statement of need should be included in the README (the intro of your paper contains good motivation, so this can essentially be used
- [ ] community guidelines (e.g. along the lines of https://github.com/MikeHeiber/Ising_OPV/blob/master/CONTRIBUTING.md)

Please let me know when you have had a chance to address these
