### 2019-02-25 791abf1
Add statement of need to README
