### 2019-02-26 #1114 @zhangning737
@lheagy taken days for reediting tests and examples, fixed.

I have submited new tests and example and add to documentation, also updated README and CONTRIBUTING. pls check it.

Thx.

Wish docs completed this time.
