### 2019-03-07 #1114 @lheagy
:wave: Hi @harpolea, would you be willing to review this submission for JOSS?
