### 2019-03-07 #1114 @harpolea
Hi @lheagy, I'm afraid I'm going to be busy travelling this month, so do not think I'd have the necessary time to review this properly (especially as I only have limited access to machines with the Intel compilers required). I also do not have much experience with finite element codes or solid mechanics, so do not think I'd be the best person to look at this.
