### 2019-03-07 #1114 @lheagy
👋 Hi @nnadeau, would you be willing to review this submission for JOSS?
