### 2019-04-30 #1328 @dvalters
Apologies for the delay to my review @lheagy, @zhaozhang, this has been due to difficulty in obtaining access to the proprietary Intel compiler required for the installation process.

Can I ask @lheagy, would this requirement be a barrier to acceptance in JOSS?

@zhaozhang , is there a way to compile this with an open-source compiler (i.e. gcc) ?
