### 2019-05-01 zhangning737/Drbsp#1 @lheagy
links to compilers in the README
(related to openjournals/joss-reviews#1328)

Are there links or recommendations you can provide as to how to obtain the compilers necessary to compile the code?
