### 2019-05-01 #1328 @lheagy
Thanks @dvalters! I have started issue zhangning737/rdbsp#1 to discuss compiler options. @zhangning737: could you please provide some pointers on where to access compilers for the code? It would be preferable if this could be achieved with an open source compiler. If this is not possible, I will check with the editorial board on any concerns for acceptance.
