### 2019-05-02 #1328 @zhangning737
@lheagy @dvalters Actually[ Intel compiler ](https://software.intel.com/en-us/parallel-studio-xe/choose-download)provide free license to open source contributors. I think just refer this repo could get a free license. Here chooses intel compiler since this software using a mkl subroutine. Of cource it could be compiled with gcc and vc community after minor changes. But I think Intel complier provides long term supports for open source community.
