### 2019-05-02 zhangning737/Drbsp#1 @zhangning737
@lheagy OK, I'll update a link, and introduce how to get a community version of intel complier.
