### 2019-05-13 34c9ff6
update how to download and install community verson of intel compiler, fix issue: https://github.com/zhangning737/rdbsp/issues/1
