### 2019-05-13 zhangning737/Drbsp#1 @zhangning737
@lheagy commit 34c9ff6 Update readme.MD and documentation with hyperlink to open accessed intel parallel studio. Pls check it. Considering to add an option for gcc in future, declaration of external fortran function has a little difference with current compiler.
