### 2019-06-10 #1328 @lheagy
:wave: Hi @jordigh, would you be willing to jump in on this review for JOSS?
