### 2019-06-30 zhangning737/Drbsp#2 @srmnitc
Compiling rdbsp throws errors
After installation of Intel Parallel studio , when ``rdbsp`` is compiled with ``Make -f Makefile_LINUX64``, it throws an error. See the attached image for the compiler message. Could be a problem with the specific setup, but in that case, more explanation in the installation section is required.
![hhh](https://user-images.githubusercontent.com/12624457/60393957-40800780-9b1d-11e9-8d56-0b9b1d7d9972.png)
