### 2019-06-30 #1328 @srmnitc
@zhangning737 @lheagy I was finally able to try compiling the code, but I was not successful. I am not sure if its a specific issue with my setup, but in any case, I am of the opinion that there should be more detailed instructions to compile successfully. I have created an [issue on the repo](https://github.com/zhangning737/rdbsp/issues/2#issue-462378692)
