### 2019-06-30 #1328 @srmnitc
@zhangning737 I have also raised an issue with citations and generally citations of books. @lheagy I am reviewing for the first time, so, do I have to take into account typos and other issues in the paper text?
