### 2019-06-30 #1328 @lheagy
@srmnitc: If you are willing to point out typos as you spot them (or create a pull request to fix them), that would be much appreciated!
