### 2019-07-01 #1328 @srmnitc
@lheagy thanks! Fixed some typos and created a pull request now.
