### 2019-07-02 #1328 @zhangning737
@srmnitc compiling error might relate to the newly update of intel compiler, I'll check it with the newest version. Thanks.
