### 2019-07-02 zhangning737/Drbsp#2 @zhangning737
@srmnitc found the problem, it should be induced by intel compiler configuration, please try to run
```
source /opt/intel/<compilers_and_libraries_version>/linux/bin/compilervars.sh intel64
```
before running
```
make -f Makefile_LINUX64
```
it will automatically configure the LD_LIBRARY_PATH to let the linker know when the libs.
Would mind get a try.
If anything wrong, pls let me know. Thanks.
