### 2019-07-02 #1328 @zhangning737
@lheagy Already pull the request. @srmnitc Please check the configuration model of Intel Compiler. I think I should add that to documentation.
