### 2019-07-02 98a29a2
remove unused files, add flex and bison to documention
