### 2019-07-02 c95c4a6
Add intel compiler configuration method to doc
