### 2019-07-02 cfc42f8
remove compiled libs
