+++ Parser.d/y.tab.h
@@ -1,217 +1,0 @@
-/* A Bison parser, made by GNU Bison 3.3.2.  */
-
-/* Bison interface for Yacc-like parsers in C
-
-   Copyright (C) 1984, 1989-1990, 2000-2015, 2018-2019 Free Software Foundation,
-   Inc.
-
-   This program is free software: you can redistribute it and/or modify
-   it under the terms of the GNU General Public License as published by
-   the Free Software Foundation, either version 3 of the License, or
-   (at your option) any later version.
-
-   This program is distributed in the hope that it will be useful,
-   but WITHOUT ANY WARRANTY; without even the implied warranty of
-   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-   GNU General Public License for more details.
-
-   You should have received a copy of the GNU General Public License
-   along with this program.  If not, see <http://www.gnu.org/licenses/>.  */
-
-/* As a special exception, you may create a larger work that contains
-   part or all of the Bison parser skeleton and distribute that work
-   under terms of your choice, so long as that work isn't itself a
-   parser generator using the skeleton or a modified version thereof
-   as a parser skeleton.  Alternatively, if you modify or redistribute
-   the parser skeleton itself, you may (at your option) remove this
-   special exception, which will cause the skeleton and the resulting
-   Bison output files to be licensed under the GNU General Public
-   License without this special exception.
-
-   This special exception was added by the Free Software Foundation in
-   version 2.2 of Bison.  */
-
-/* Undocumented macros, especially those whose name start with YY_,
-   are private implementation details.  Do not rely on them.  */
-
-#ifndef YY_YY_Y_TAB_H_INCLUDED
-# define YY_YY_Y_TAB_H_INCLUDED
-/* Debug traces.  */
-#ifndef YYDEBUG
-# define YYDEBUG 0
-#endif
-#if YYDEBUG
-extern int yydebug;
-#endif
-
-/* Token type.  */
-#ifndef YYTOKENTYPE
-# define YYTOKENTYPE
-  enum yytokentype
-  {
-    IntConstant = 258,
-    DblConstant = 259,
-    NODE = 260,
-    NewLine = 261,
-    ELEMENT = 262,
-    ZERO = 263,
-    STRNAME = 264,
-    END = 265,
-    MATERIAL = 266,
-    DISP = 267,
-    FORCE = 268,
-    IVEL = 269,
-    IACC = 270,
-    ISTR = 271,
-    JAUMA = 272,
-    GRAV = 273,
-    ATT = 274,
-    CURVE = 275,
-    TERM = 276,
-    PRESSURE = 277,
-    FIXED = 278,
-    TIMESTEP = 279,
-    DYNA = 280,
-    SOLV = 281,
-    OUTPUT = 282,
-    NOUTPUTID = 283,
-    EOUTPUTID = 284,
-    HISTORY = 285,
-    HISOUTPUT = 286,
-    HOURGLASS = 287,
-    STATUSFILE = 288,
-    TITLE = 289,
-    REMOVE = 290,
-    BULK_VISCOSITY = 291,
-    HISNODE = 292,
-    HISELEM = 293,
-    DEBUGFLG = 294,
-    CLEARDVFLG = 295,
-    LOCALDAMP = 296,
[…]
