### 2019-07-08 #1328 @srmnitc
@lheagy just finished looking at it. Most issues are now fixed - installation and tests work perfectly except for some minor changes which I have kept open as an issue. The DOIs on the paper are incomplete, so I have kept the issue open.
