### 2019-07-08 #1328 @srmnitc
@lheagy @zhangning737 I still think a better documentation is required - the existing one is complete, but inaccessible from a non specialist perspective. A detailed documentation with example usage would be a recommendation for the future.
