### 2019-07-08 #1328 @srmnitc
@lheagy @zhangning737 There are only two issues holding back my review - 1) the DOIs of references are not complete 2) There is no release version of github
