### 2019-07-09 5fa8dfd
Updated the reference of paper
