+++ paper/paper.bib
@@ -4,7 +4,6 @@
   isbn={9781138032286},
   lccn={2016058563},
   series={ISRM book series},
-  url={https://books.google.de/books?id=kXv8vQAACAAJ},
   year={2017},
   publisher={Taylor \& Francis Group, an informa business}
 }
@@ -16,7 +15,6 @@ journal = {Tunnelling and Underground Space Technology},
 keywords = { Dynamic failure constitutive model, Inversion analysis, Large deformation,Deep rock mass},
 pages = {47--53},
 title = {{Computational method of large deformation and its application in deep mining tunnel}},
-url = {http://www.sciencedirect.com/science/article/pii/S0886779815001054},
 volume = {50},
 year = {2015}
 }
@@ -29,7 +27,6 @@ month = {mar},
 number = {2},
 pages = {131--149},
 title = {{Stabilization of DAEs and invariant manifolds}},
-url = {https://doi.org/10.1007/s002110050020},
 volume = {67},
 year = {1994}
 }
@@ -40,7 +37,6 @@ isbn = {9780080525693},
 publisher = {Elsevier Science},
 series = {The Mechanics of Constitutive Modeling},
 title = {{The Mechanics of Constitutive Modeling}},
-url = {https://books.google.de/books?id=t2a4tl6uRVYC},
 year = {2005}
 }
 
@@ -50,7 +46,6 @@ year = {2005}
   isbn={9780080951355},
   lccn={2014397935},
   series={The Finite Element Method},
-  url={https://books.google.de/books?id=7UL5Ls9hOF8C},
   year={2013},
   publisher={Elsevier Science}
 }
@@ -61,8 +56,8 @@ journal = {International Journal for Numerical Methods in Engineering},
 number = {5},
 pages = {679--706},
 title = {{A uniform strain hexahedron and quadrilateral with orthogonal hourglass control}},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nme.1620170504},
-volume = {17}
+volume = {17},
+year = {1981}
 }
 
 @article{VALANIS1990353,
@@ -72,7 +67,6 @@ journal = {International Journal of Plasticity},
 number = {3},
 pages = {353--367},
 title = {{Back stress and Jaumann rates in finite plasticity}},
-url = {http://www.sciencedirect.com/science/article/pii/0749641990900072},
 volume = {6},
 year = {1990}
 }
@@ -82,7 +76,6 @@ author = {Cundall, P A},
 journal = {Analytical and Computational Methods in Engineering Rock Mechanics},
 pages = {129--163},
 title = {{Distinct element models of rock and soil structure}},
-url = {https://ci.nii.ac.jp/naid/10003877143/en/},
 year = {1987}
 }
 
