+++ paper/paper.bib
@@ -5,7 +5,8 @@
   lccn={2016058563},
   series={ISRM book series},
   year={2017},
-  publisher={Taylor \& Francis Group, an informa business}
+  doi={10.1201/9781315391304},
+  publisher={CRC Press}
 }
 
 @article{WANG201547,
@@ -14,6 +15,7 @@ issn = {0886-7798},
 journal = {Tunnelling and Underground Space Technology},
 keywords = { Dynamic failure constitutive model, Inversion analysis, Large deformation,Deep rock mass},
 pages = {47--53},
+doi = {10.1016/j.tust.2015.06.006},
 title = {{Computational method of large deformation and its application in deep mining tunnel}},
 volume = {50},
 year = {2015}
@@ -26,6 +28,7 @@ journal = {Numerische Mathematik},
 month = {mar},
 number = {2},
 pages = {131--149},
+doi = {10.1007/s002110050020},
 title = {{Stabilization of DAEs and invariant manifolds}},
 volume = {67},
 year = {1994}
@@ -35,6 +38,7 @@ year = {1994}
 author = {Ottosen, N S and Ristinmaa, M},
 isbn = {9780080525693},
 publisher = {Elsevier Science},
+doi = {10.1016/B978-0-08-044606-6.X5000-0},
 series = {The Mechanics of Constitutive Modeling},
 title = {{The Mechanics of Constitutive Modeling}},
 year = {2005}
@@ -46,6 +50,7 @@ year = {2005}
   isbn={9780080951355},
   lccn={2014397935},
   series={The Finite Element Method},
+  doi = {10.1016/C2009-0-24909-9},
   year={2013},
   publisher={Elsevier Science}
 }
@@ -56,6 +61,7 @@ journal = {International Journal for Numerical Methods in Engineering},
 number = {5},
 pages = {679--706},
 title = {{A uniform strain hexahedron and quadrilateral with orthogonal hourglass control}},
+doi = {10.1002/nme.1620170504},
 volume = {17},
 year = {1981}
 }
@@ -67,6 +73,7 @@ journal = {International Journal of Plasticity},
 number = {3},
 pages = {353--367},
 title = {{Back stress and Jaumann rates in finite plasticity}},
+doi = {10.1016/0749-6419(90)90007-2},
 volume = {6},
 year = {1990}
 }
@@ -75,6 +82,7 @@ year = {1990}
 author = {Cundall, P A},
 journal = {Analytical and Computational Methods in Engineering Rock Mechanics},
 pages = {129--163},
+doi = {10.1002/eqe.4290160314},
 title = {{Distinct element models of rock and soil structure}},
 year = {1987}
 }
@@ -82,15 +90,15 @@ year = {1987}
 @online{Kitware,
   author = {Kitware},
   title = {paraview.org},
-  year = 2018,
+  year = 2019,
   url = {https://www.paraview.org},
-  urldate = {2018-11-30}
+  urldate = {2019-1-30}
 }
 
 @online{Tecplot,
   author = {Tecplot},
   title = {tecplot.com},
-  year = 2018,
+  year = 2019,
   url = {https://www.tecplot.com},
-  urldate = {2018-11-30}
+  urldate = {2019-1-30}
 }
