### 2019-07-10 620ca8b
Update paper.md
minor grammatical suggestions to the paper in the JOSS review. Please double check my edits and ensure I have not mis-represented any statements in the article.
