### 2019-07-10 #1328 @lheagy
👋 Hi @zhangning737, I started pr zhangning737/rdbsp#6 with a few minor grammatical updates to the paper. Please take a look when you have a chance.
