### 2019-07-11 #1328 @lheagy
Excellent, thanks @zhangning737, can you please take a  look at the article proof and make sure everything is in order? Once you are happy with it, please generate a new software release and archive it on [zenodo](http://zenodo.org) or similar and set the title and author of the archive to be the same as that in your paper. After that, please post the doi here and we can proceed with publication.
