### 2019-07-15 #1328 @zhangning737
@lheagy Just changed 3 words in the last paragraph of paper.

Already upload to ``` zenodo ``` and the DOI is:

```10.5281/zenodo.3335406```

Is the DOI alright? seems a little bit odd...
If there anything else need to be done, please let me know. Thanks...
