### 2019-07-23 #1328 @lheagy
👋 @openjournals/joss-eics, this submission is ready to be published! One reviewer did not fully complete the review, so please let me know if you need me to jump in and explicitly check off those items as a third reviewer. Thanks!
