### 2019-07-23 #1328 @kyleniemeyer
Hi @zhangning737, it looks like you deleted the JOSS paper file in your latest commit (https://github.com/zhangning737/rdbsp/commit/52742007231dba84846a82a38bf398f9dafa75f7). Can you add that back?
