### 2019-07-24 343649d
Add paper back
