### 2019-07-24 #1328 @kyleniemeyer
Thanks @zhangning737. Could you next merge the pull request I just submitted, which made a few minor edits to the paper? https://github.com/zhangning737/rdbsp/pull/7
