### 2019-07-25 #1328 @BarryDeCicco
I do not want to be involved with JOSS any more.  I would really appreciate it if somebody would remove me from this.  I've running at a rate of 50 e-mails this morning alone.
Barry

    On Thursday, July 25, 2019, 12:20:10 PM EDT, whedon <notifications@github.com> wrote:

Check final proof 👉 openjournals/joss-papers#858

If the paper PDF and Crossref deposit XML look good in openjournals/joss-papers#858, then you can now move forward with accepting the submission by compiling again with the flag deposit=true e.g.
