### 2019-07-25 #1328 @kyleniemeyer
Hi @BarryDeCicco, you are likely subscribed to all notifications from the `joss-reviews` repo, which happens by default when you are added as a reviewer. (And would lead to quite a lot of unnecessary notifications.) You can fix this by going to the "Watch" dropdown at the top of the page, and changing to "not watching":

![](https://cloud.githubusercontent.com/assets/4483/20250593/64d7ce48-a9de-11e6-9225-d3dfb3e48e68.png)
