### 2019-07-25 #1328 @kyleniemeyer
Congrats @zhangning737 on your paper's publication in JOSS! Thanks to @dvalters and @srmnitc for reviewing, and @lheagy for editing!
