### 2019-07-25 #1328 @BarryDeCicco
Thank you, Kyle!  I’ve set myself to ‘ignoring’.
Barry
    On Thursday, July 25, 2019, 12:36:04 PM EDT, Kyle Niemeyer <notifications@github.com> wrote:

Hi @BarryDeCicco, you are likely subscribed to all notifications from the joss-reviews repo, which happens by default when you are added as a reviewer. (And would lead to quite a lot of unnecessary notifications.) You can fix this by going to the "Watch" dropdown at the top of the page, and changing to "not watching":
